Moved menu item stuff in sub package
