+++ src/ui/graphical/menuItem/MenuItemDelegate.java
@@ -1,4 +1,4 @@
-package ui.graphical;
+package ui.graphical.menuItem;
 
 public interface MenuItemDelegate {
 	public void performAction(MenuItemType type);
