+++ src/ui/graphical/menuItem/MenuItemType.java
@@ -1,4 +1,4 @@
-package ui.graphical;
+package ui.graphical.menuItem;
 
 public enum MenuItemType {
 	New,
